+++ bin/logstash.bat
@@ -11,9 +11,9 @@ set first_arg=%1
 setlocal EnableDelayedExpansion
 if "!first_arg:~0,1!" equ "-" (
   if "%VENDORED_JRUBY%" == "" (
-    %RUBYCMD% "%LS_HOME%\lib\bootstrap\environment.rb" "logstash\runner.rb" agent %*
+    %RUBYCMD% "%LS_HOME%\lib\bootstrap\environment.rb" "logstash\runner.rb" %*
   ) else (
-    %JRUBY_BIN% %jruby_opts% "%LS_HOME%\lib\bootstrap\environment.rb" "logstash\runner.rb" agent %*
+    %JRUBY_BIN% %jruby_opts% "%LS_HOME%\lib\bootstrap\environment.rb" "logstash\runner.rb" %*
   )
 ) else (
   if "%VENDORED_JRUBY%" == "" (
