+++ logstash-core/lib/logstash/runner.rb
@@ -3,22 +3,82 @@ Thread.abort_on_exception = true
 Encoding.default_external = Encoding::UTF_8
 $DEBUGLIST = (ENV["DEBUG"] || "").split(",")
 
+require "clamp" # gem 'clamp'
+require "net/http"
 require "logstash/environment"
 
 LogStash::Environment.load_locale!
 
 require "logstash/namespace"
-require "logstash/program"
+require "logstash/agent"
 
-class LogStash::Runner
-  include LogStash::Program
+class LogStash::Runner < Clamp::Command
 
-  def main(args)
+  DEFAULT_INPUT = "input { stdin { type => stdin } }"
+  DEFAULT_OUTPUT = "output { stdout { codec => rubydebug } }"
+
+  option ["-f", "--config"], "CONFIG_PATH",
+    I18n.t("logstash.runner.flag.config"),
+    :attribute_name => :config_path
+
+  option "-e", "CONFIG_STRING",
+    I18n.t("logstash.runner.flag.config-string",
+           :default_input => DEFAULT_INPUT, :default_output => DEFAULT_OUTPUT),
+    :default => "", :attribute_name => :config_string
+
+  option ["-w", "--filterworkers"], "COUNT",
+    I18n.t("logstash.runner.flag.filterworkers"),
+    :attribute_name => :filter_workers,
+    :default => LogStash::Config::CpuCoreStrategy.fifty_percent, &:to_i
+
+  option ["-l", "--log"], "FILE",
+    I18n.t("logstash.runner.flag.log"),
+    :attribute_name => :log_file
+
+  # Old support for the '-v' flag'
+  option "-v", :flag,
+    I18n.t("logstash.runner.flag.verbosity"),
+    :attribute_name => :verbosity, :multivalued => true
+
+  option "--quiet", :flag, I18n.t("logstash.runner.flag.quiet")
+  option "--verbose", :flag, I18n.t("logstash.runner.flag.verbose")
+  option "--debug", :flag, I18n.t("logstash.runner.flag.debug")
+
+  option ["-V", "--version"], :flag,
+    I18n.t("logstash.runner.flag.version")
+
+  option ["-p", "--pluginpath"] , "PATH",
+    I18n.t("logstash.runner.flag.pluginpath"),
+    :multivalued => true,
+    :attribute_name => :plugin_paths
+
+  option ["-t", "--configtest"], :flag,
+    I18n.t("logstash.runner.flag.configtest"),
+    :attribute_name => :config_test
+
+  option "--[no-]allow-unsafe-shutdown", :flag,
+    I18n.t("logstash.agent.flag.unsafe_shutdown"),
+    :attribute_name => :unsafe_shutdown,
+    :default => false
+
+  option ["-i", "--interactive"], "SHELL",
+    I18n.t("logstash.runner.flag.rubyshell"),
+    :attribute_name => :ruby_shell
+
+  attr_reader :agent
+
+  def initialize(*args)
+    @agent = LogStash::Agent.new
+    super(*args)
+  end
+
+  def execute
     require "logstash/util"
     require "logstash/util/java_version"
-    require "stud/trap"
     require "stud/task"
-    @startup_interruption_trap = Stud::trap("INT") { puts "Interrupted"; exit 0 }
+    require "cabin" # gem 'cabin'
+
+    @logger = Cabin::Channel.get(LogStash)
 
     LogStash::Util::set_thread_name(self.class.name)
 
@@ -30,95 +90,256 @@ class LogStash::Runner
     # Print a warning to STDERR for bad java versions
     LogStash::Util::JavaVersion.warn_on_bad_java_version
 
-    Stud::untrap("INT", @startup_interruption_trap)
+    LogStash::ShutdownController.unsafe_shutdown = unsafe_shutdown?
+    LogStash::ShutdownController.logger = @logger
 
-    task = run(args)
-    exit(task.wait)
+    configure
+
+    if version?
+      show_version
+      return 0
+    end
+
+    return start_shell(@ruby_shell, binding) if @ruby_shell
+
+    if @config_string.nil? && @config_path.nil?
+      fail(I18n.t("logstash.runner.missing-configuration"))
+    end
+
+    @agent.logger = @logger
+
+    config_string = format_config(@config_path, @config_string)
+
+    @agent.add_pipeline("base", config_string, :filter_workers => filter_workers)
+
+    if config_test?
+      puts "Configuration OK"
+    else
+      task = Stud::Task.new { @agent.execute }
+      return task.wait
+    end
+
+  rescue LoadError => e
+    fail("Configuration problem.")
+  rescue LogStash::ConfigurationError => e
+    @logger.fatal I18n.t("logstash.error", :error => e)
+    if !config_test?
+      @logger.warn I18n.t("logstash.runner.configtest-flag-information")
+    end
+    show_short_help
+    return 1
+  rescue => e
+    @logger.fatal I18n.t("oops", :error => e)
+    @logger.debug e.backtrace if $DEBUGLIST.include?("stacktrace")
+  ensure
+    @log_fd.close if @log_fd
   end # def self.main
 
-  def run(args)
-    command = args.shift
-    commands = {
-      "version" => lambda do
-        require "logstash/agent"
-        agent_args = ["--version"]
-        if args.include?("--verbose")
-          agent_args << "--verbose"
-        end
-        return LogStash::Agent.run($0, agent_args)
-      end,
-      "irb" => lambda do
-        require "irb"
-        return IRB.start(__FILE__)
-      end,
-      "pry" => lambda do
-        require "pry"
-        return binding.pry
-      end,
-      "docgen" => lambda do
-        require 'docs/asciidocgen'
-        opts = OptionParser.new
-        settings = {}
-        opts.on("-o DIR", "--output DIR",
-          "Directory to output to; optional. If not specified,"\
-          "we write to stdout.") do |val|
-          settings[:output] = val
-        end
-        args = opts.parse(ARGV)
-        docs = LogStashConfigAsciiDocGenerator.new
-        args.each do |arg|
-          docs.generate(arg, settings)
-        end
-        return 0
-      end,
-      "agent" => lambda do
-        require "logstash/agent"
-        # Hack up a runner
-        agent = LogStash::Agent.new("/bin/logstash agent", $0)
-        begin
-          agent.parse(args)
-        rescue Clamp::HelpWanted => e
-          show_help(e.command)
-          return 0
-        rescue Clamp::UsageError => e
-          # If 'too many arguments' then give the arguments to
-          # the next command. Otherwise it's a real error.
-          raise if e.message != "too many arguments"
-          remaining = agent.remaining_arguments
-        end
+  def show_version
+    show_version_logstash
 
-        return agent.execute
-      end
-    } # commands
-
-    if commands.include?(command)
-      return Stud::Task.new { commands[command].call }
-    else
-      if command.nil?
-        $stderr.puts "No command given"
-      else
-        if !%w(--help -h help).include?(command)
-          # Emit 'no such command' if it's not someone asking for help.
-          $stderr.puts "No such command #{command.inspect}"
-        end
-      end
-      $stderr.puts %q[
-Usage: logstash <command> [command args]
-Run a command with the --help flag to see the arguments.
-For example: logstash agent --help
-
-Available commands:
-  agent - runs the logstash agent
-  version - emits version info about this logstash
-]
-      #$stderr.puts commands.keys.map { |s| "  #{s}" }.join("\n")
-      return Stud::Task.new { 1 }
+    if [:info, :debug].include?(verbosity?) || debug? || verbose?
+      show_version_ruby
+      show_version_java if LogStash::Environment.jruby?
+      show_gems if [:debug].include?(verbosity?) || debug?
     end
-  end # def run
+  end # def show_version
 
-  private
+  def show_version_logstash
+    require "logstash/version"
+    puts "logstash #{LOGSTASH_VERSION}"
+  end # def show_version_logstash
 
-  def show_help(command)
-    puts command.help
+  def show_version_ruby
+    puts RUBY_DESCRIPTION
+  end # def show_version_ruby
+
+  def show_version_java
+    properties = java.lang.System.getProperties
+    puts "java #{properties["java.version"]} (#{properties["java.vendor"]})"
+    puts "jvm #{properties["java.vm.name"]} / #{properties["java.vm.version"]}"
+  end # def show_version_java
+
+  def show_gems
+    require "rubygems"
+    Gem::Specification.each do |spec|
+      puts "gem #{spec.name} #{spec.version}"
+    end
+  end # def show_gems
+
+  # Do any start-time configuration.
+  #
+  # Log file stuff, plugin path checking, etc.
+  def configure
+    configure_logging(log_file)
+    configure_plugin_paths(plugin_paths)
+  end # def configure
+
+  # add the given paths for ungemified/bare plugins lookups
+  # @param paths [String, Array<String>] plugins path string or list of path strings to add
+  def configure_plugin_paths(paths)
+    Array(paths).each do |path|
+      fail(I18n.t("logstash.runner.configuration.plugin_path_missing", :path => path)) unless File.directory?(path)
+      LogStash::Environment.add_plugin_path(path)
+    end
+  end
+
+  # Point logging at a specific path.
+  def configure_logging(path)
+    # Set with the -v (or -vv...) flag
+    if quiet?
+      @logger.level = :error
+    elsif verbose?
+      @logger.level = :info
+    elsif debug?
+      @logger.level = :debug
+    else
+      # Old support for the -v and -vv stuff.
+      if verbosity? && verbosity?.any?
+        # this is an array with length of how many times the flag is given
+        if verbosity?.length == 1
+          @logger.warn("The -v flag is deprecated and will be removed in a future release. You should use --verbose instead.")
+          @logger.level = :info
+        else
+          @logger.warn("The -vv flag is deprecated and will be removed in a future release. You should use --debug instead.")
+          @logger.level = :debug
+        end
+      else
+        @logger.level = :warn
+      end
+    end
+
+    if log_file
+      # TODO(sissel): Implement file output/rotation in Cabin.
+      # TODO(sissel): Catch exceptions, report sane errors.
+      begin
+        @log_fd.close if @log_fd
+        @log_fd = File.new(path, "a")
+      rescue => e
+        fail(I18n.t("logstash.runner.configuration.log_file_failed",
+                    :path => path, :error => e))
+      end
+
+      @logger.subscribe(STDOUT, :level => :fatal)
+      @logger.subscribe(@log_fd)
+      @logger.terminal "Sending logstash logs to #{path}."
+    else
+      @logger.subscribe(STDOUT)
+    end
+
+    # TODO(sissel): redirect stdout/stderr to the log as well
+    # http://jira.codehaus.org/browse/JRUBY-7003
+  end # def configure_logging
+
+  def format_config(config_path, config_string)
+    config_string = config_string.to_s
+    if config_path
+      # Append the config string.
+      # This allows users to provide both -f and -e flags. The combination
+      # is rare, but useful for debugging.
+      config_string = config_string + load_config(config_path)
+    else
+      # include a default stdin input if no inputs given
+      if config_string !~ /input *{/
+        config_string += DEFAULT_INPUT
+      end
+      # include a default stdout output if no outputs given
+      if config_string !~ /output *{/
+        config_string += DEFAULT_OUTPUT
+      end
+    end
+    config_string
+  end
+
+  def load_config(path)
+    begin
+      uri = URI.parse(path)
+
+      case uri.scheme
+      when nil then
+        local_config(path)
+      when /http/ then
+        fetch_config(uri)
+      when "file" then
+        local_config(uri.path)
+      else
+        fail(I18n.t("logstash.runner.configuration.scheme-not-supported", :path => path))
+      end
+    rescue URI::InvalidURIError
+      # fallback for windows.
+      # if the parsing of the file failed we assume we can reach it locally.
+      # some relative path on windows arent parsed correctly (.\logstash.conf)
+      local_config(path)
+    end
+  end
+
+  def local_config(path)
+    path = File.expand_path(path)
+    path = File.join(path, "*") if File.directory?(path)
+
+    if Dir.glob(path).length == 0
+      fail(I18n.t("logstash.runner.configuration.file-not-found", :path => path))
+    end
+
+    config = ""
+    encoding_issue_files = []
+    Dir.glob(path).sort.each do |file|
+      next unless File.file?(file)
+      if file.match(/~$/)
+        @logger.debug("NOT reading config file because it is a temp file", :config_file => file)
+        next
+      end
+      @logger.debug("Reading config file", :config_file => file)
+      cfg = File.read(file)
+      if !cfg.ascii_only? && !cfg.valid_encoding?
+        encoding_issue_files << file
+      end
+      config << cfg + "\n"
+      if config_test?
+        @logger.debug? && @logger.debug("\nThe following is the content of a file", :config_file => file.to_s)
+        @logger.debug? && @logger.debug("\n" + cfg + "\n\n")
+      end
+    end
+    if (encoding_issue_files.any?)
+      fail("The following config files contains non-ascii characters but are not UTF-8 encoded #{encoding_issue_files}")
+    end
+    if config_test?
+      @logger.debug? && @logger.debug("\nThe following is the merged configuration")
+      @logger.debug? && @logger.debug("\n" + config + "\n\n")
+    end
+    return config
+  end # def load_config
+
+  def fetch_config(uri)
+    begin
+      Net::HTTP.get(uri) + "\n"
+    rescue Exception => e
+      fail(I18n.t("logstash.runner.configuration.fetch-failed", :path => uri.to_s, :message => e.message))
+    end
+  end
+
+  # Emit a failure message and abort.
+  def fail(message)
+    raise LogStash::ConfigurationError, message
+  end # def fail
+
+  def show_short_help
+    puts I18n.t("logstash.runner.short-help")
+  end
+
+  def start_shell(shell, start_binding)
+    case shell
+    when "pry"
+      require 'pry'
+      start_binding.pry
+    when "irb"
+      require 'irb'
+      ARGV.clear
+      # TODO: set binding to this instance of Runner
+      # currently bugged as per https://github.com/jruby/jruby/issues/384
+      IRB.start(__FILE__)
+    else
+      fail(I18n.t("logstash.runner.invalid-shell"))
+    end
   end
 end # class LogStash::Runner
